+++ src/main/java/animatch/app/api/configuration/security/AutenticacaoProvider.java
@@ -1,0 +1,41 @@
+package animatch.app.api.configuration.security;
+
+import animatch.app.service.usuario.autenticacao.AutenticacaoService;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+public class AutenticacaoProvider  implements AuthenticationProvider {
+    private final AutenticacaoService usuarioAutorizacaoService;
+    private final PasswordEncoder passwordEncoder;
+
+    public AutenticacaoProvider(AutenticacaoService usuarioAutorizacaoService, PasswordEncoder passwordEncoder) {
+        this.usuarioAutorizacaoService = usuarioAutorizacaoService;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Override
+    public Authentication authenticate(final Authentication authentication) throws AuthenticationException {
+        final String username = authentication.getName();
+        final String password = authentication.getCredentials().toString();
+        System.out.println("USUARIO: "+ username + "\nCREDENCIAIS: "+password);
+
+        UserDetails userDetails = this.usuarioAutorizacaoService.loadUserByUsername(username);
+        System.out.println("DETAILS: "+userDetails.getPassword());
+
+        if (this.passwordEncoder.matches(password, userDetails.getPassword())) {
+            return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+        } else {
+            throw new BadCredentialsException("Usuário ou Senha inválidos");
+        }
+    }
+
+    @Override
+    public boolean supports(final Class<?> authentication) {
+        return authentication.equals(UsernamePasswordAuthenticationToken.class);
+    }
+}
