+++ src/main/java/animatch/app/domain/comentario/Comentario.java
@@ -1,0 +1,51 @@
+package animatch.app.domain.comentario;
+
+import animatch.app.domain.topico.Topico;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+
+@Entity
+public class Comentario {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    @NotBlank
+    private String texto;
+    @ManyToOne
+    private Topico topico;
+    @ManyToOne
+    private Comentario comentarioPai;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getTexto() {
+        return texto;
+    }
+
+    public void setTexto(String texto) {
+        this.texto = texto;
+    }
+
+    public Topico getTopico() {
+        return topico;
+    }
+
+    public void setTopico(Topico topico) {
+        this.topico = topico;
+    }
+
+    public Comentario getComentario() {
+        return comentarioPai;
+    }
+
+    public void setComentario(Comentario comentario) {
+        this.comentarioPai = comentario;
+    }
+
+}
