+++ src/main/java/animatch/app/domain/lista/Lista.java
@@ -1,0 +1,42 @@
+package animatch.app.domain.lista;
+
+import animatch.app.domain.usuario.Usuario;
+import jakarta.persistence.Id;
+import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.validation.constraints.NotNull;
+
+@Entity
+public class Lista {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    @NotNull
+    @ManyToOne
+    private Usuario userId;
+    private String name;
+
+    public Lista() {
+    }
+    public Lista(String name) {
+        this.name = name;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public Usuario getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Usuario userId) {
+        this.userId = userId;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
