+++ src/main/java/animatch/app/domain/topico/repository/TopicoRepository.java
@@ -1,0 +1,8 @@
+package animatch.app.domain.topico.repository;
+
+import animatch.app.domain.topico.Topico;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TopicoRepository extends JpaRepository<Topico,Integer> {
+    Topico findTopicoById(int id);
+}
