+++ src/main/java/animatch/app/domain/usuario/Usuario.java
@@ -1,0 +1,102 @@
+package animatch.app.domain.usuario;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.*;
+import org.springframework.validation.annotation.Validated;
+
+import java.time.LocalDate;
+
+@Validated
+@Entity
+public class Usuario {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Schema(description = "Identificador único", example = "1")
+    private int id;
+    @Schema(description = "Nome do usuário", example = "fulano de tal")
+    private String name;
+    @Email
+//    @Pattern(
+//        regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,}$"
+//    )
+    @Schema(description = "E-mail do usuário", example = "usuario@email.com")
+    private String email;
+    @Schema(description = "Senha do usuario", example = "senhasecreta")
+    private String password;
+    @Schema(description = "Link para imagem de perfil", example = "")
+    private String profileImage;
+    @Schema(description = "Link para imagem de capa do perfil", example = "")
+    private String coverImage;
+    @Schema(description = "Data de nascimento", example = "2000-01-01")
+    @Past
+    private LocalDate nascimento;
+    private boolean status = true;
+
+    public Usuario() {
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public String getProfileImage() {
+        return profileImage;
+    }
+
+    public String getCoverImage() {
+        return coverImage;
+    }
+
+    public LocalDate getNascimento() {
+        return nascimento;
+    }
+
+    public boolean isStatus() {
+        return status;
+    }
+
+    public void setStatus(boolean status) {
+        this.status = status;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setProfileImage(String profileImage) {
+        this.profileImage = profileImage;
+    }
+
+    public void setCoverImage(String coverImage) {
+        this.coverImage = coverImage;
+    }
+
+    public void setNascimento(LocalDate nascimento) {
+        this.nascimento = nascimento;
+    }
+}
