+++ src/main/java/animatch/app/dto/AnimeInfoDTO.java
@@ -1,6 +1,6 @@
 package animatch.app.dto;
 
-import animatch.app.model.Anime;
+import animatch.app.domain.anime.Anime;
 
 public class AnimeInfoDTO {
     private Anime anime;
