--- src/main/java/animatch/app/dto/UsuarioDTO.java
@@ -1,27 +1,0 @@
-package animatch.app.dto;
-
-public class UsuarioDTO {
-    private String email;
-    private String password;
-
-    public UsuarioDTO(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
