--- src/main/java/animatch/app/model/Lista.java
@@ -1,41 +1,0 @@
-package animatch.app.model;
-
-import jakarta.persistence.Id;
-import jakarta.persistence.*;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.validation.constraints.NotNull;
-
-@Entity
-public class Lista {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    @NotNull
-    @ManyToOne
-    private Usuario userId;
-    private String name;
-
-    public Lista() {
-    }
-    public Lista(String name) {
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public Usuario getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Usuario userId) {
-        this.userId = userId;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
