--- src/main/java/animatch/app/model/Usuario.java
@@ -1,66 +1,0 @@
-package animatch.app.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.*;
-import org.springframework.validation.annotation.Validated;
-
-import java.time.LocalDate;
-
-@Validated
-@Entity
-public class Usuario {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private String name;
-    @Email
-//    @Pattern(
-//        regexp = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,}$"
-//    )
-    private String email;
-    private String password;
-    private String profileImage;
-    private String coverImage;
-    @Past
-    private LocalDate nascimento;
-    private boolean status = true;
-
-    public Usuario() {
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public String getProfileImage() {
-        return profileImage;
-    }
-
-    public String getCoverImage() {
-        return coverImage;
-    }
-
-    public LocalDate getNascimento() {
-        return nascimento;
-    }
-
-    public boolean isStatus() {
-        return status;
-    }
-
-    public void setStatus(boolean status) {
-        this.status = status;
-    }
-}
