--- src/main/java/animatch/app/repository/AnimeRepository.java
@@ -1,7 +1,0 @@
-package animatch.app.repository;
-
-import animatch.app.model.Anime;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AnimeRepository extends JpaRepository<Anime,Integer> {
-}
