--- src/main/java/animatch/app/repository/ListaRepository.java
@@ -1,28 +1,0 @@
-package animatch.app.repository;
-
-import animatch.app.model.Usuario;
-import animatch.app.dto.ListaInfoDTO;
-import org.springframework.data.jpa.repository.JpaRepository;
-import animatch.app.model.Lista;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-
-public interface ListaRepository extends JpaRepository<Lista,Integer> {
-
-    List<Lista> findAllByUserId(Usuario userId);
-
-    @Query("""
-            select new animatch.app.dto.ListaInfoDTO(l.id, l.name)
-            from Lista l where l.userId in :userId
-            """)
-    List<ListaInfoDTO> findAllListaInfoByUserId(Usuario userId);
-
-    @Query("""
-            select new animatch.app.dto.ListaInfoDTO(l.id, l.name)
-            from Lista l
-            """)
-    List<ListaInfoDTO> findAllInfo();
-
-    Lista findListaById(int id);
-}
