--- src/main/java/animatch/app/repository/TopicoRepository.java
@@ -1,8 +1,0 @@
-package animatch.app.repository;
-
-import animatch.app.model.Topico;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TopicoRepository extends JpaRepository<Topico,Integer> {
-    Topico findTopicoById(int id);
-}
