+++ src/main/java/animatch/app/service/usuario/autenticacao/AutenticacaoService.java
@@ -1,0 +1,25 @@
+package animatch.app.service.usuario.autenticacao;
+
+import animatch.app.domain.usuario.Usuario;
+import animatch.app.domain.usuario.repository.UsuarioRepository;
+import animatch.app.service.usuario.autenticacao.dto.UsuarioDetalhesDTO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+@Service
+public class AutenticacaoService implements UserDetailsService {
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Usuario usuario = usuarioRepository.findByEmail(username);
+        if (usuario == null) {
+            throw new UsernameNotFoundException(String.format("usuario: %s nao encontrado", username));
+        }
+        return new UsuarioDetalhesDTO(usuario);
+    }
+}
