+++ src/main/java/animatch/app/service/usuario/dto/UsuarioLoginMapper.java
@@ -1,0 +1,10 @@
+package animatch.app.service.usuario.dto;
+
+public class UsuarioLoginMapper {
+    public static UsuarioLoginDTO of(UsuarioLoginDTO usuarioLoginDTO){
+        final UsuarioLoginDTO usuario = new UsuarioLoginDTO();
+        usuario.setPassword(usuarioLoginDTO.getPassword());
+        usuario.setEmail(usuarioLoginDTO.getEmail());
+        return usuario;
+    }
+}
