+++ src/main/resources/application.properties
@@ -10,3 +10,14 @@ spring.jpa.properties.hibernate.format_sql=true
 # permite o acesso ao h2 console: localhost:8080/h2-console
 spring.h2.console.enabled=true
 spring.datasource.url=jdbc:h2:mem:animes
+
+#server.servlet.context-path=/anime-match
+# validade do token
+jwt.validity=3600000
+# palavra passe do token (segredo) necessita de no minimo 32 caracteres
+jwt.secret=RXhpc3RlIHVtYSB0ZW9yawEgcXVLIGRpeiBxdWUsIHNlIHVtIGRpYSBhbGd16W0gZGVzY29icmlyIGV4YXRhb\
+  WVudGUgcGFyYSBxdWUgc2VydmUgbyBVbml2ZXJzbyBlIHBvciBxdWUgZWxlIGVzd0EgYXFlaSwgZWxLIGRLc2FwYXJlY2Vy4\
+  SBpbnN0YW50YW51YW1lbnRlIGUgc2Vy4SBzdwJzdGl0delkbyBwb3IgYWxnbyBhaW5kYSBtYWlzIGVzdHJhbmhvIGUgaW5le\
+  HBsaWPhdmVsLLBFeGlzdGUgdW1hIHNlZ3VuZGEgdGVvcm1hIHF1ZSBkaXogcXVlIGlzc28gauEgYWNvbnRlY2V1L14u
+
+Spring.jpa.defer-datasource-initialization=true
